Mail List action bar improvements

+++ Wino.Mail/Controls/Advanced/WinoListView.cs
@@ -120,7 +120,7 @@ namespace Wino.Controls.Advanced
             {
                 args.Handled = true;
 
-                ItemDeletedCommand?.Execute((int)MailOperation.SoftDelete);
+                ItemDeletedCommand?.Execute(MailOperation.SoftDelete);
             }
         }
